+++ app/(anon)/recruitments/[id]/_components/recruitmentContent.module.scss
