+++ app/(anon)/recruitments/[id]/recruitmentDetail.module.scss
