+++ app/(anon)/recruitments/_components/recruitmentsItem/recruitmentsItem.module.scss
