+++ app/(anon)/recruitments/_components/recruitmentsItem/recruitmentsItem.tsx
@@ -1,8 +1,8 @@
 import Badge from "@/components/badge/badge";
 
-import styles from "./RecruitmentsItem.module.scss";
+import styles from "./recruitmentsItem.module.scss";
 
-import RecruitmentsTag from "../RecruitmentsTag/RecruitmentsTag";
+import RecruitmentsTag from "../recruitmentsTag/recruitmentsTag";
 
 import { FaHeart, FaEye, FaComment } from "react-icons/fa";
