+++ app/(anon)/recruitments/_components/recruitmentsTag/recruitmentsTag.module.scss
