+++ app/user/layout.tsx
@@ -2,8 +2,8 @@
 
 import type { ReactNode } from "react";
 
-import Header from "@/components/header/Header";
-import Footer from "@/components/footer/Footer";
+import Header from "@/components/header/header";
+import Footer from "@/components/footer/footer";
 
 import styles from "./layout.module.scss";
