+++ app/user/projects/[id]/_components/applicantDetails.tsx
@@ -1,4 +1,4 @@
-import styles from "../ProjectDetail.module.scss";
+import styles from "../projectDetail.module.scss";
 
 import type { Applicant } from "./projectData";
