+++ app/user/projects/[id]/projectDetail.module.scss
