+++ app/user/projects/myParticipate/page.tsx
@@ -1,7 +1,7 @@
 // 마이페이지 - 참여한 프로젝트
 import styles from "./myParticipate.module.scss";
 
-import MyProjectItem from "../_components/myProjectItem/MyProjectItem";
+import MyProjectItem from "../_components/myProjectItem/myProjectItem";
 
 export default function Page() {
   return (
