+++ app/user/recruitments/_components/myApplyStatusItem/myApplyStatusItem.module.scss
