+++ app/user/recruitments/myComment/page.tsx
@@ -1,7 +1,7 @@
 // 마이페이지 - 댓글 쓴 모집글
 import styles from "./myComment.module.scss";
 
-import MyRecruitmentItem from "../_components/myRecruitmentItem/MyRecruitmentItem";
+import MyRecruitmentItem from "../_components/myRecruitmentItem/myRecruitmentItem";
 
 export default function Page() {
   return (
