+++ components/inputField/inputField.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 
-import styles from "./InputField.module.scss";
+import styles from "./inputField.module.scss";
 
 import type { IconType } from "react-icons";
